@@ -3,8 +3,8 @@ name = "running-stats"
 version = "0.1.3"
 description = "Library for computing linear statistics."
 readme = "README.md"
-authors = [{ name = "Kristoffer Andersson", email = "kod.kristoff@gmail.com" }]
-requires-python = ">=3.9"
+authors = [{ name = "Språkbanken Text", email = "sb-info@svenska.gu.se" }]
+requires-python = ">=3.10"
 dependencies = ["attrs>=25.1.0"]
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -14,11 +14,11 @@ classifiers = [
     "Operating System :: Unix",
     "Operating System :: POSIX",
     # "Operating System :: Microsoft :: Windows",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "Programming Language :: Python :: Implementation :: CPython",
     # "uncomment if you test on these interpreters:",
     # "Programming Language :: Python :: Implementation :: IronPython",
@@ -29,13 +29,13 @@ classifiers = [
 license = { text = "MIT" }
 
 [project.urls]
-Homepage = "https://github.com/kod-kristoff/running-stats-py"
+Homepage = "https://spraakbanken.gu.se"
 Repository = "https://github.com/kod-kristoff/running-stats-py"
-"Bug Tracker" = "https://github.com/kod-kristoff/running-stats-py/issues"
+Issues = "https://github.com/kod-kristoff/running-stats-py/issues"
 
 [build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+requires = ["uv_build>=0.9.6,<0.10.0"]
+build-backend = "uv_build"
 
 [dependency-groups]
 dev = ["mypy>=1.15.0", "pytest>=8.3.4", "pytest-cov>=6.0.0", "ruff>=0.9.7"]
